--- data/test/tf_peaks_TEST_sparse_Remap.h5
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:22255164dcc03400203c5af6d257f8df8b6d334f6f635eaf4dd58c6f867c4d28
-size 154599439
