--- explainn/__init__.py
@@ -1,11 +1,0 @@
-"""
-ExplaiNN
-interpretable and transparent neural networks for genomics
-"""
-
-from .utils import tools
-from .interpretation import interpretation
-from .train import train, test
-from .models import networks
-
-__version__ = "0.1.5"
